langgraph/pyproject.toml
[project]
name = "langgraph-demo"
version = "0.1.0"
description = "LangGraph Demo"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.2.12",
    "click>=8.2.1",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.3",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]

[tool.setuptools]
packages = ["app"] 
